+++ Editor/Tests/Engine/YieldTest.cs
@@ -17,7 +17,7 @@ namespace Exodrifter.Rumor.Test.Engine
 			var yield = new ForAdvance();
 			Assert.False(yield.Finished);
 
-			yield.OnUpdate(0);
+			yield.OnUpdate(null, 0);
 			Assert.False(yield.Finished);
 
 			yield.OnAdvance();
@@ -33,16 +33,70 @@ namespace Exodrifter.Rumor.Test.Engine
 		[Test]
 		public void YieldChoice()
 		{
-			var yield = new ForChoice();
+			// Choose 1
+			var yield = new ForChoice(1, 0, 0);
 			Assert.False(yield.Finished);
 
-			yield.OnUpdate(0);
+			yield.OnUpdate(null, 0);
 			Assert.False(yield.Finished);
 
 			yield.OnAdvance();
 			Assert.False(yield.Finished);
 
 			yield.OnChoice();
+			Assert.True(yield.Finished);
+
+			yield.OnChoice();
+			Assert.True(yield.Finished);
+
+			yield.OnUpdate(null, 0);
+			Assert.True(yield.Finished);
+
+			yield.OnAdvance();
+			Assert.True(yield.Finished);
+
+			// Multiple
+			yield = new ForChoice(2, 0, 0);
+			Assert.False(yield.Finished);
+
+			yield.OnUpdate(null, 0);
+			Assert.False(yield.Finished);
+
+			yield.OnAdvance();
+			Assert.False(yield.Finished);
+
+			yield.OnChoice();
+			Assert.False(yield.Finished);
+
+			yield.OnChoice();
+			Assert.True(yield.Finished);
+
+			yield.OnChoice();
+			Assert.True(yield.Finished);
+
+			yield.OnUpdate(null, 0);
+			Assert.True(yield.Finished);
+
+			yield.OnAdvance();
+			Assert.True(yield.Finished);
+
+			// Timed
+			yield = new ForChoice(2, 2, 0);
+			Assert.False(yield.Finished);
+
+			yield.OnUpdate(null, 1);
+			Assert.False(yield.Finished);
+
+			yield.OnAdvance();
+			Assert.False(yield.Finished);
+
+			yield.OnUpdate(null, 1);
+			Assert.True(yield.Finished);
+
+			yield.OnUpdate(null, 1);
+			Assert.True(yield.Finished);
+
+			yield.OnAdvance();
 			Assert.True(yield.Finished);
 
 			yield.OnChoice();
@@ -64,13 +118,13 @@ namespace Exodrifter.Rumor.Test.Engine
 			yield.OnChoice();
 			Assert.False(yield.Finished);
 
-			yield.OnUpdate(0);
+			yield.OnUpdate(null, 0);
 			Assert.False(yield.Finished);
 
-			yield.OnUpdate(1);
+			yield.OnUpdate(null, 1);
 			Assert.True(yield.Finished);
 
-			yield.OnUpdate(1);
+			yield.OnUpdate(null, 1);
 			Assert.True(yield.Finished);
 		}
 	}